+++ docs/changelog.rst
@@ -10,6 +10,9 @@ v1.2.0
 Bug Fixes
   * :issue:`352`: Remove duplicate entry from user list.
 
+Miscellaneous
+  * :issue:`355`: Bump package versions
+
 
 ******
 v1.1.0