+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 v1.2.0
 ******
 
+Features
+  * :issue:`354`: Add Know Me user information to accessor response.
+
 Bug Fixes
   * :issue:`352`: Remove duplicate entry from user list.